requirement.txt
streamlit
langchain-openai==0.3.14
langchain-core==0.3.54
langchain-tools==0.1.34
GitPython==3.1.44
python-dotenv==1.1.0
